+++ examples/repeat/repeat.go
@@ -12,7 +12,7 @@ func main() {
 	cli.AutoHelp(true)
 
 	// you can use `--times 3` or `-t 3`
-	t := cli.FlagInt("time", "t", "Number of times to repeat the string").AsRequired()
+	t := cli.FlagInt("times", "t", "Number of times to repeat the string").AsRequired()
 
 	// message is a variadic positional argument, so you can provide multiple
 	// values for it, and they will be joined together with spaces to form the